% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkSeleniumServer}
\alias{checkSeleniumServer}
\title{Check for Server binary (baloise version)}
\usage{
checkSeleniumServer(dir = NULL, update = FALSE,
  proxyUrl = "webproxy.balgroupit.com",
  selName = "selenium-server-standalone.jar")
}
\arguments{
\item{selName}{The name of the Selenium Server File.}

\item{update}{A boolean indicating whether to update the binary if it is present.}

\item{proxyUrl}{}

\item{selName}{}
}
\description{
\code{checkSeleniumServer}
A utility function to check if the Selenium Server stanalone binary is present.
}
\section{Detail}{
 The downloads for the Selenium project can be found at http://selenium-release.storage.googleapis.com/index.html. This convience function downloads the standalone server and places it in the RSelenium package directory bin folder by default.
}
\examples{
\dontrun{
checkSeleniumServer(update=F)
}
}

