% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chromeDriver}
\alias{chromeDriver}
\title{Use Chrome Profiles}
\usage{
chromeDriver(use_profile = FALSE, profile_name = "Default",
  profile_dir = "Default", internal_testing = FALSE)
}
\arguments{
\item{use_profile}{}

\item{profile_name}{Selects directory of profile to associate with the first browser launched.}

\item{profile_dir}{Specifies the user data directory, which is where the browser will look for all of its state.}

\item{internal_testing}{}
}
\description{
\code{chromeDriver}
A utility function for easy and automated access to the Chrome driver.
}
\section{Detail}{

The profile directory is contained in the user directory and by default is named "Default"
}

