% Generated by roxygen2 (4.0.1): do not edit by hand
\name{chromeDriver}
\alias{chromeDriver}
\title{Use Chrome Profiles}
\usage{
chromeDriver(useProfile = FALSE, profileName = "Default",
  profileDir = "Default", internalTesting = FALSE)
}
\arguments{
\item{useProfile}{}

\item{profileDir}{Specifies the user data directory, which is where the browser will look for all of its state.}

\item{profileName}{Selects directory of profile to associate with the first browser launched.}

\item{internalTesting}{}
}
\description{
\code{chromeDriver}
A utility function for easy and automated access to the Chrome driver.
}
\section{Detail}{

The profile directory is contained in the user directory and by default is named "Default"
}

