% Generated by roxygen2 (4.0.2): do not edit by hand
\name{firefoxDriver}
\alias{firefoxDriver}
\title{Use Firefox Profiles}
\usage{
firefoxDriver(use_profile = FALSE, profile_name = "default",
  profile_dir = "default", base_encrypton = TRUE)
}
\arguments{
\item{use_profile}{}

\item{profile_name}{Selects directory of profile to associate with the first browser launched.}

\item{profile_dir}{Specifies the user data directory, which is where the browser will look for all of its state.}

\item{base_encrypton}{}
}
\description{
\code{firefoxDriver}
A utility function for easy and automated access to the firefox driver.
}
\section{Detail}{

The profile directory is contained in the user directory and by default is named "Default"
}

