% Generated by roxygen2 (4.0.2): do not edit by hand
\name{firefoxDriver}
\alias{firefoxDriver}
\title{Use Firefox Profiles}
\usage{
firefoxDriver(useProfile = FALSE, profileName = "default",
  profileDir = "default", baseEncrypton = TRUE)
}
\arguments{
\item{useProfile}{}

\item{profileName}{Selects directory of profile to associate with the first browser launched.}

\item{profileDir}{Specifies the user data directory, which is where the browser will look for all of its state.}

\item{baseEncrypton}{}
}
\description{
\code{firefoxDriver}
A utility function for easy and automated access to the firefox driver.
}
\section{Detail}{

The profile directory is contained in the user directory and by default is named "Default"
}

